issue #86 - Does _generic.translation still need hardcoded "Apply", "Cancel" ... responses?
Confirmation from Malav is that these hardcoded strings are not needed.

To create valid JSON, each term supplied by the user is prepended with
a ',' separator. This works fine except that it needs a first
key:value pair hardcoded in the output. So I replaced the multiple
unused translation strings with a single identifying key:value.